little refactor

+++ src/main/java/sk/stuba/fei/uim/oop/tiles/StraightPipe.java
@@ -3,7 +3,6 @@ package sk.stuba.fei.uim.oop.tiles;
 import sk.stuba.fei.uim.oop.board.Direction;
 
 import java.awt.*;
-import java.awt.geom.AffineTransform;
 
 public class StraightPipe extends Tile{
     public StraightPipe() {
@@ -24,7 +23,7 @@ public class StraightPipe extends Tile{
             this.highlight = false;
         }
         g2d.setStroke(new BasicStroke(4));
-        g2d.drawRect(0, (int)this.getHeight()/2-this.getHeight()/6, this.getWidth(), (int)this.getWidth()/3);
+        g2d.drawRect(0, this.getHeight() /2-this.getHeight()/6, this.getWidth(), this.getWidth() /3);
         g2d.setStroke(new BasicStroke(1));
         if(water){
             g2d.setColor(Color.BLUE);
@@ -32,7 +31,7 @@ public class StraightPipe extends Tile{
         else {
             g2d.setColor(Color.GRAY);
         }
-        g2d.fillRect(0, (int)this.getHeight()/2-this.getHeight()/6, this.getWidth(), (int)this.getWidth()/3);
+        g2d.fillRect(0, this.getHeight() /2-this.getHeight()/6, this.getWidth(), this.getWidth() /3);
         g2d.setColor(Color.BLACK);
         g2d.dispose();
     }
